Microservices/DashboardMS/requirements.txt
Flask==3.0.2
paho-mqtt
numpy
influxdb_client
python-dotenv
requests